fix llm runtime issues

+++ intel_extension_for_transformers/neural_chat/tests/optimization/test_optimization.py
@@ -59,7 +59,7 @@ class TestChatbotBuilder(unittest.TestCase):
     def test_build_chatbot_with_llm_runtime(self):
         loading_config = LoadingModelConfig(use_llm_runtime=True)
         config = PipelineConfig(model_name_or_path="/models/opt-125m",
-            optimization_config=WeightOnlyQuantConfig(),
+            optimization_config=WeightOnlyQuantConfig(compute_dtype="int8", weight_dtype="int8"),
             loading_config=loading_config
         )
         chatbot = build_chatbot(config)
